Create README.md
This is my very first project as a software developer. It tells you a little bit of who i am, what i love doing and my favorite meals. This project was done with the basic tools of HTML and CSS. Thanks to Alkimi Space for making my dream gradually become my reality.